requirements.txt
# requirements.txt

absl-py==2.1.0
asciimatics==1.15.0
asciinema==2.4.0
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
build==1.0.3
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.0
contourpy==1.0.7
cryptography==42.0.5
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.8
Django==4.2.7
dm-tree==0.1.8
dnssec==0.0.0
etils==1.8.0
executing==2.0.1
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.39.4
fsspec==2024.3.1
func-timeout==4.3.5
gast==0.5.4
google-api-core==2.15.0
google-api-python-client==2.115.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
grpcio==1.63.0
h5py==3.11.0
httplib2==0.22.0
idna==3.4
importlib_resources==6.4.0
intel-openmp==2021.4.0
ipykernel==6.26.0
ipython==8.18.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.0
jupyter_core==5.5.0
keras==3.3.3
kiwisolver==1.4.4
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
nest-asyncio==1.5.8
networkx==3.3
numpy==1.26.2
oauthlib==3.2.2
opt-einsum==3.3.0
optree==0.11.0
packaging==23.1
pandas==2.1.4
parso==0.8.3
Pillow==9.5.0
platformdirs==4.0.0
promise==2.3
prompt-toolkit==3.0.41
protobuf==4.25.3
psutil==5.9.6
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycparser==2.21
pyfiglet==1.0.2
Pygments==2.17.2
pyparsing==3.0.9
pyproject_hooks==1.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
PyX==0.16
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
scapy==2.5.0
schedule==1.2.1
scipy==1.10.1
shellingham==1.5.4
six==1.16.0
sounddevice==0.4.6
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.12
tbb==2021.12.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow-datasets==4.9.4
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.15.0
termcolor==2.4.0
toml==0.10.2
torch==2.3.0+cpu
torchaudio==2.3.0
torchvision==0.18.0+cpu
tornado==6.3.3
tqdm==4.66.4
traitlets==5.13.0
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
virtualenv==20.25.0
wcwidth==0.2.12
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.18.1
